requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
pystray>=0.19.5
keyboard>=0.13.5
numpy>=1.24.0
pyperclip>=1.8.2
pywin32>=306
